Refactor by splitting the files
